TeNe.jl Documentation

TeNe.contract — Method

contract(x, y, cix, ciy, [conjx=false, conjy=false]; kwargs...)

Contract tensors x and y across dimensions cix and ciy, and returns it as z.

Arguments

- `x`: first tensor to contract.
- `y': second tensor to contract.
- `cix`: the dimensions of the first tensor to contract.
- `ciy`: the dimensions of the second tensor to contract.
- `conjx::Bool=false`: Take the complex conjugate of argument x?
- `conjy::Bool=false`: Take the complex conjugate of argument y?

Optional Keyword Arguments

- `tocache::Bool=true`: store the result in the second level of the cache?
- `sublevel=:auto`: if stored in cache, at which sublevel? :auto finds non-aliased memory

Examples

julia> x = randn(ComplexF64, 2, 3, 4);
julia> y = randn(ComplexF64, 3, 5, 6);
julia> z = contract(x, y, 2, 1);
julia> size(z)
(2, 4, 5, 6)

julia> x = randn(ComplexF64, 2, 3, 4, 5);
julia> y = randn(ComplexF64, 6, 5, 2, 7);
julia> z = contract(x, y, (1, 4), (3, 2));
julia> size(z)
(3, 4, 6, 7)